+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.2.0]
+
+- add Environment variables explorer
+- fix Docker verification issue
+- improve FluentCI Explorer UI/UX
+
 ## [0.1.2]
 
 - verify all fluentci requirements